requirements.txt
annotated-types~=0.5.0
anyio~=3.7.1
fastapi~=0.103.1 
passlib[bcrypt]~=1.7.3
psycopg2-binary~=2.9.7  
pydantic~=2.3.0
pydantic-settings~=2.0.3 
python-jose~=3.3.0 
python-multipart~=0.0.6  
sqlalchemy~=2.0.20  
uvicorn~=0.23.2